Refactor merchant GraphQL API
