+++ backend/tilavarauspalvelu/api/graphql/schema.py
@@ -38,7 +38,6 @@ from .mutations import (
     EquipmentUpdateMutation,
     PurposeCreateMutation,
     PurposeUpdateMutation,
-    RefreshOrderMutation,
     RejectAllApplicationOptionsMutation,
     RejectAllSectionOptionsMutation,
     ReservationAdjustTimeMutation,
@@ -120,7 +119,7 @@ from .queries import (
     UnitNode,
     UserNode,
 )
-from .types.merchants.permissions import PaymentOrderPermission
+from .types.payment_order.permissions import PaymentOrderPermission
 
 if TYPE_CHECKING:
     import datetime
@@ -375,9 +374,6 @@ class Mutation(graphene.ObjectType):
     deny_reservation_series = ReservationSeriesDenyMutation.Field()
     change_reservation_series_access_code = ReservationSeriesChangeAccessCodeMutation.Field()
     repair_reservation_series_access_code = ReservationSeriesRepairAccessCodeMutation.Field()
-    #
-    # Orders
-    refresh_order = RefreshOrderMutation.Field()
     #
     # User
     update_current_user = CurrentUserUpdateMutation.Field()
