--- backend/tilavarauspalvelu/api/graphql/types/merchants/serializers.py
@@ -1,13 +1,0 @@
-from __future__ import annotations
-
-from rest_framework import serializers
-
-
-class RefreshOrderInputSerializer(serializers.Serializer):
-    order_uuid = serializers.UUIDField(required=True)
-
-
-class RefreshOrderOutputSerializer(serializers.Serializer):
-    order_uuid = serializers.UUIDField()
-    status = serializers.CharField()
-    reservation_pk = serializers.IntegerField()
